OS independent install.py script.
git-svn-id: http://svn.wxwidgets.org/svn/wx/wxPython@51680 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ 3rdParty/FloatCanvas/install.py
@@ -1,32 +1,49 @@
 #!/usr/bin/env python2.4
-
-# A script that installs the floatcanvas package in the wxPython lib It
-# also installs the demo into the wxPython demo directory, if you set
-# the path for that correctly NOTE: I can't figure out how to get
-# distutils to put it into a subpackage of an existing package.
-#   Now I do: package_dir = {"wx/lib": ""} or something like that should work.
-
-DemoDir = "/home/cbarker/wxPython-2.6.3.0/demo"
+'''!
+   A script that installs the floatcanvas package in the wxPython lib It
+   also installs the demo into the wxPython demo directory, if you set
+   the path for that correctly.
+   
+   CB -- Now use the os.path module to make the install script OS independent.
+   
+   Issues:
+   - SVN's folder .svn doesn't delete with shutil.
+   - Demo folder must be hard coded.
+'''
 
 import sys, os, shutil
+import os.path
 
 import wx
 
-InstallDir = wx.__path__[0]+"/lib/floatcanvas"
+# The installation Directory
+InstallDir = os.path.join(wx.__path__[0],"lib","floatcanvas")
+
+# The Demo Directory (Please Set correctly or leave blank.)
+Demodir = ''
+
+# ##########################################################################
+
+# Remove the old package if its there the package there.
+if os.path.isdir(InstallDir):
+    shutil.rmtree(InstallDir)
+else:
+    print "FloatCanvas failed to install or doesn't exist.\nHave you removed the .svn subfolder fro mthe previous install?"
 
 
-print "Copying floatcanvas to:", InstallDir
-# copy the package there.
-shutil.rmtree(InstallDir)
-shutil.copytree("./floatcanvas",InstallDir)
+# Copy-in the flatcavas package from the current working directory.
+try:
+    print "Copying floatcanvas to:", InstallDir
+    shutil.copytree(".%sfloatcanvas"%(os.path.sep),InstallDir)
+    # Python raise an exception for one of the file in the folder below. I don't know why. Manual delete works fine.
+    print 'If you installed from a SVN package, please remove the .svn folder.'
+except:
+    print "Failed to copy to install directory, do you have permission to do so?"
 
-#print SITE_PACKAGES
-#SITE_PACKAGES=/usr/lib/python2.3/site-packages
-#cp -r floatcanvas/ $SITE_PACKAGES/wxPython/lib/
 
-# Now install the demo:
-
-print "Copying the demo to:", DemoDir
-shutil.copyfile("./FloatCanvasDemo.py", os.path.join(DemoDir,"FloatCanvas.py") )
+# Install the demo if the path is specified.
+if Demodir:
+    print "Copying the demo to:", DemoDir
+    shutil.copyfile(".%sFloatCanvasDemo.py"%(os.path.sep), os.path.join(DemoDir,"FloatCanvas.py") )
